* arts-1-12-92
* src/m_rte.cc (1475):

Implicit conversion from Index to Numeric throws a warning in gcc
4.3.3 because the value might change. Added explicit cast.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6132 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/m_rte.cc
@@ -1472,7 +1472,7 @@ void iyClearskyStandard(
     { 
       iy_aux.resize( 2, stokes_dim, nf ); 
       //
-      iy_aux( 1, joker, joker ) = i_background;
+      iy_aux( 1, joker, joker ) = (Numeric)i_background;
     }    
   //
   if( j_analytical_do  &&  diy_dx.nelem() == 0 )
